moved stress code to separate submodule
