+++ src/napari_stress/_stress/charts_SPB.py
