+++ src/napari_stress/_stress/lebedev_info_SPB.py
