+++ src/napari_stress/_stress/manifold_SPB.py
