+++ src/napari_stress/_stress/sph_func_SPB.py
